requirements.txt
fastmcp
pydantic
uvicorn
starlette
google-adk
streamlit
litellm
